+++ Changelog.rst
@@ -5,6 +5,11 @@ version 1.8.7.0
 **2020-09-??**
 
 * New keyword parameter to `cfdm.Field.set_data`: ``inplace``
+* New class: `cfdm.core.meta.DocstringRewriteMeta`
+* New method: `cfdm.Field._docstring_special_substitutions`
+* New method: `cfdm.Field._docstring_substitutions`
+* New method: `cfdm.Field._docstring_package_depth`
+* New method: `cfdm.Field._docstring_method_exclusions`
 * Fixed bug that caused failures when reading or writing a dataset
   that contains multiple geometry containers
   (https://github.com/NCAS-CMS/cfdm/issues/65)